Fix typos in entries and curate materials
The tutorial for creating vector fields was removed
since the assets section should primarily focus on
assets ready for use in the engine or for tools to
aid with development. While a tutorial is helpful,
there are a wide variety of topics that could be
addressed which would cause an explosion in the
number of entries for each topic.